backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic==2.11.7
pydantic-settings==2.10.1
pandas==2.3.2
python-dotenv==1.1.1
python-multipart==0.0.20
python-docx==1.2.0
pypdf==5.9.0
google-generativeai==0.8.5
pytest==8.4.1
httpx==0.28.1
